Hot Potato
Hot Potato is a Python program that assists job-searching developers in keeping their GitHub heat map as active as they want it to appear. Realistically GitHub's heat map is not a metric that hiring managers should be using to vet potential employees, but realistically some of them will. So job searching developers should prepare and use this app to keep their heat map hot!